requirements.txt
astroid==2.0.2
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
defusedxml==0.5.0
Django==2.1
django-allauth==0.36.0
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lime==0.1.1.31
matplotlib==2.2.2
mccabe==0.6.1
mkl-fft==1.0.5
mkl-random==1.0.1
networkx==2.1
numpy==1.15.0
oauthlib==2.1.0
pandas==0.23.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.0
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
PyWavelets==0.5.2
reportlab==3.5.4
requests==2.19.1
requests-oauthlib==1.0.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
toolz==0.9.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
wrapt==1.10.11
